refactor(FirebaseActionHandler): change file type to TSX and fix login path
